# docs/swagger.yaml
basePath: /api/v1
consumes:
- application/json
definitions:
...
  models.FollowerResponse:
    properties:
      avatar:
        $ref: '#/definitions/models.Photo'
      first_name:
        type: string
      followed_at:
        type: string
      full_name:
        type: string
      is_followed_by:
        type: boolean
      is_following:
        description: Relationship status (relative to authenticated user)
        type: boolean
      last_name:
        type: string
      province:
        type: string
      user_id:
        type: string
    type: object
  models.FollowingResponse:
    properties:
      avatar:
        $ref: '#/definitions/models.Photo'
      first_name:
        type: string
      following_since:
        type: string
      full_name:
        type: string
      is_followed_by:
        type: boolean
      is_following:
        description: Relationship status (relative to authenticated user)
        type: boolean
      last_name:
        type: string
      province:
        type: string
      user_id:
        type: string
    type: object
...
  models.Photo:
    properties:
      height:
        type: integer
      mime_type:
        type: string
      name:
        type: string
      size:
        type: integer
      url:
        type: string
      width:
        type: integer
    type: object
...
  models.RelationshipStatus:
    properties:
      has_blocked_me:
        type: boolean
      is_blocked:
        type: boolean
      is_followed_by:
        type: boolean
      is_following:
        type: boolean
    type: object
...
  utils.Response:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
host: localhost:8080
info:
  contact:
    email: support@hamsaya.app
    name: API Support
    url: https://hamsaya.app/support
  description: A production-ready Go backend for a social media mobile application
    with posts, business profiles, marketplace, real-time chat, and location services.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://hamsaya.app/terms/
  title: Hamsaya Backend API
  version: "1.0"
paths:
...
  /users/{user_id}/block:
    delete:
      description: Unblock a previously blocked user
      parameters:
      - description: User ID to unblock
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Unblock a user
      tags:
      - relationships
    post:
      description: Block another user
      parameters:
      - description: User ID to block
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Block a user
      tags:
      - relationships
  /users/{user_id}/follow:
    delete:
      description: Unfollow a user that you're currently following
      parameters:
      - description: User ID to unfollow
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Unfollow a user
      tags:
      - relationships
    post:
      description: Follow another user
      parameters:
      - description: User ID to follow
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Follow a user
      tags:
      - relationships
  /users/{user_id}/followers:
    get:
      description: Get a list of users following the specified user
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      - default: 20
        description: Limit
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.FollowerResponse'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Get followers
      tags:
      - relationships
  /users/{user_id}/following:
    get:
      description: Get a list of users that the specified user is following
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      - default: 20
        description: Limit
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.FollowingResponse'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Get following
      tags:
      - relationships
  /users/{user_id}/relationship:
    get:
      description: Get the relationship status between authenticated user and another
        user
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.RelationshipStatus'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Get relationship status
      tags:
      - relationships
...
produces:
- application/json
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"